+++ src/commands/applyFilterToTreeView.ts
@@ -1,14 +1,21 @@
 import { QuickPickItem, window } from 'vscode';
 import { $config, $state } from '../extension';
+import { FILTER_CONSTANTS } from '../filter';
 import { tasksView, updateTasksTreeView } from '../treeViewProviders/treeViews';
-import { VscodeContext } from '../types';
-import { setContext } from '../utils/vscodeUtils';
+import { VscodeContext, setContext } from '../vscodeContext';
+import { setGlobalSate } from '../vscodeGlobalState';
 
-export function applyFilterToTreeView() {
+export function applyFilterToTreeViewCommand() {
 	const quickPick = window.createQuickPick();
-	quickPick.items = $config.savedFilters.map(fl => ({
-		label: fl.title,
-	}) as QuickPickItem);
+	quickPick.items = [
+		...$config.savedFilters.map(savedFilter => ({
+			label: savedFilter.title,
+		}) as QuickPickItem),
+		...Object.keys(FILTER_CONSTANTS).map(savedFilterKey => ({
+			label: savedFilterKey,
+			detail: FILTER_CONSTANTS[savedFilterKey as keyof typeof FILTER_CONSTANTS],
+		}) as QuickPickItem),
+	];
 	let value: string | undefined;
 	let selected: string | undefined;
 	quickPick.onDidChangeValue(e => {
@@ -21,7 +28,13 @@ export function applyFilterToTreeView() {
 	quickPick.onDidAccept(() => {
 		let filterStr;
 		if (selected) {
+			// Saved filter
 			filterStr = $config.savedFilters.find(fl => fl.title === selected)?.filter;
+			// Filter Constant
+			if (!filterStr) {
+				// @ts-ignore
+				filterStr = FILTER_CONSTANTS[selected];
+			}
 		} else {
 			filterStr = value;
 		}
@@ -30,9 +43,18 @@ export function applyFilterToTreeView() {
 		if (!filterStr || !filterStr.length) {
 			return;
 		}
-		tasksView.description = filterStr;
-		setContext(VscodeContext.FilterActive, true);
-		$state.taskTreeViewFilterValue = filterStr;
-		updateTasksTreeView();
+		applyTreeViewFilter(filterStr);
 	});
+}
+
+export function applyTreeViewFilter(filterStr: string | undefined): void {
+	if (filterStr) {
+		tasksView.description = `Filter: ${filterStr}`;
+	} else {
+		tasksView.description = undefined;
+	}
+	setContext(VscodeContext.FilterActive, Boolean(filterStr));
+	$state.taskTreeViewFilterValue = filterStr;
+	setGlobalSate('tasksTreeViewFilterValue', filterStr);
+	updateTasksTreeView();
 }