+++ src/commands/showDefaultFileSetting.ts
@@ -1,6 +1,0 @@
-import { Constants } from '../constants';
-import { openSettingsGuiAt } from '../utils/vscodeUtils';
-
-export function showDefaultFileSetting() {
-	openSettingsGuiAt(Constants.DefaultFileSetting);
-}