+++ src/events.ts
@@ -5,12 +5,11 @@ import { getNextFewTasks } from './commands/getFewNextTasks';
 import { Constants } from './constants';
 import { disposeDecorations, doUpdateEditorDecorations } from './decorations';
 import { resetAllRecurringTasks } from './documentActions';
-import { $config, $state, counterStatusBar, mainStatusBar, updateLastVisitGlobalState, updateState } from './extension';
+import { $config, $state, updateLastVisitGlobalState, updateState } from './extension';
 import { clearDiagnostics, updateDiagnostic } from './languageFeatures/diagnostics';
-import { updateAllTreeViews } from './treeViewProviders/treeViews';
-import { VscodeContext } from './types';
+import { updateAllTreeViews, updateArchivedTasks } from './treeViewProviders/treeViews';
 import { getDocumentForDefaultFile } from './utils/extensionUtils';
-import { setContext } from './utils/vscodeUtils';
+import { setContext, VscodeContext } from './vscodeContext';
 
 let changeTextDocumentDisposable: Disposable | undefined;
 let changeActiveTextEditorDisposable: Disposable | undefined;
@@ -21,15 +20,16 @@ let changeActiveTextEditorDisposable: Disposable | undefined;
  * This event can be fired multiple times very quickly 5-20ms interval.
  */
 export async function onChangeActiveTextEditor(editor: TextEditor | undefined): Promise<void> {
-	if ($state.theRightFileOpened) {
+	if ($state.activeEditorMatchesActivatePattern) {
 		deactivateEditorFeatures();
 	}
-	if (editor && isTheRightFileName(editor)) {
+	if (editor && isActiveFileMatchesActivatePattern(editor)) {
 		$state.activeDocument = editor.document;
 		$state.activeDocumentTabSize = typeof editor.options.tabSize === 'number' ? editor.options.tabSize : $config.tabSize;
 		await updateEverything(editor);
 		activateEditorFeatures(editor);
 		await setContext(VscodeContext.IsActive, true);
+		$state.isActiveFileTheArchiveFile = isActiveFileTheArchiveFile(editor);
 
 		const needReset = checkIfNeedResetRecurringTasks(editor.document.uri.toString());
 		if (needReset) {
@@ -40,10 +40,14 @@ export async function onChangeActiveTextEditor(editor: TextEditor | undefined): 
 	} else {
 		$state.activeDocument = await getDocumentForDefaultFile();
 		$state.activeDocumentTabSize = $config.tabSize;
-		$state.theRightFileOpened = false;
+		$state.activeEditorMatchesActivatePattern = false;
+		$state.isActiveFileTheArchiveFile = false;
 		await updateEverything();
 		await setContext(VscodeContext.IsActive, false);
 	}
+}
+function isActiveFileTheArchiveFile(editor: TextEditor): boolean {
+	return editor.document.uri.fsPath === $config.defaultArchiveFile;
 }
 /**
  * Only run reset all recurring tasks when needed (first open file in a day)
@@ -72,14 +76,18 @@ export function checkIfNeedResetRecurringTasks(filePath: string): {lastVisit: Da
  */
 export function onChangeTextDocument(e: TextDocumentChangeEvent) {
 	const activeTextEditor = window.activeTextEditor;
-	if (activeTextEditor && $state.theRightFileOpened) {
+	if (activeTextEditor && $state.activeEditorMatchesActivatePattern) {
 		updateEverything(activeTextEditor);
+
+		if ($state.isActiveFileTheArchiveFile) {
+			updateArchivedTasks();
+		}
 	}
 }
 /**
  * Match Uri of editor against a glob specified by user.
  */
-export function isTheRightFileName(editor: TextEditor): boolean {
+export function isActiveFileMatchesActivatePattern(editor: TextEditor): boolean {
 	return languages.match({
 		pattern: $config.activatePattern,
 	},	editor.document) !== 0;
@@ -88,16 +96,16 @@ export function isTheRightFileName(editor: TextEditor): boolean {
  * Activate document text change event listener.
  */
 export function activateEditorFeatures(editor: TextEditor) {
-	$state.theRightFileOpened = true;
+	$state.activeEditorMatchesActivatePattern = true;
 	changeTextDocumentDisposable = workspace.onDidChangeTextDocument(onChangeTextDocument);
-	counterStatusBar.show();
+	$state.progressStatusBar.show();
 }
 /**
  * Deactivate document text change event listener.
  */
 export function deactivateEditorFeatures() {
 	changeTextDocumentDisposable?.dispose();
-	counterStatusBar.hide();
+	$state.progressStatusBar.hide();
 }
 export function disposeEditorDisposables() {
 	disposeDecorations();
@@ -123,13 +131,13 @@ export function disposeActiveEditorChange() {
  */
 export const updateEverything = throttle(async (editor?: TextEditor) => {
 	await updateState();
-	if (editor && isTheRightFileName(editor)) {
+	if (editor && isActiveFileMatchesActivatePattern(editor)) {
 		doUpdateEditorDecorations(editor);
-		counterStatusBar.update($state.tasks);
+		$state.progressStatusBar.update($state.tasks);
 		updateDiagnostic(editor, $state.tasksAsTree);
 	} else {
 		clearDiagnostics();
 	}
-	mainStatusBar.update(getNextFewTasks());
+	$state.mainStatusBar.update(getNextFewTasks());
 	updateAllTreeViews();
 }, Constants.ThrottleEverything);