+++ src/treeViewProviders/tagProvider.ts
@@ -1,8 +1,8 @@
 import { Event, EventEmitter, TreeDataProvider, TreeItem, TreeItemCollapsibleState } from 'vscode';
-import { getTaskHoverMd } from '../languageFeatures/getTaskHover';
 import { TheTask } from '../TheTask';
+import { getTasksHoverMd } from '../languageFeatures/getTaskHover';
 import { ItemForProvider } from '../types';
-import { tasksToTreeItems, TaskTreeItem } from './taskProvider';
+import { TaskTreeItem, tasksToTreeItems } from './taskProvider';
 
 export class TagTreeItem extends TreeItem {
 	readonly collapsibleState = TreeItemCollapsibleState.Collapsed;
@@ -35,7 +35,7 @@ export class TagProvider implements TreeDataProvider<TagTreeItem | TaskTreeItem>
 	 */
 	resolveTreeItem(item: TagTreeItem | TaskTreeItem, el: TagTreeItem | TaskTreeItem) {
 		if (el instanceof TaskTreeItem) {
-			el.tooltip = getTaskHoverMd(el.task);
+			el.tooltip = getTasksHoverMd([el.task]);
 			return el;
 		}
 		return undefined;