+++ src/treeViewProviders/taskProvider.ts
@@ -2,9 +2,9 @@ import { Command, Event, EventEmitter, ThemeColor, ThemeIcon, TreeDataProvider, 
 import { TheTask } from '../TheTask';
 import { CommandId } from '../commands';
 import { $config } from '../extension';
-import { getTaskHoverMd } from '../languageFeatures/getTaskHover';
+import { getTasksHoverMd } from '../languageFeatures/getTaskHover';
 import { defaultSortTasks } from '../sort';
-import { SortNestedTasks } from '../types';
+import { IsDue, SortNestedTasks } from '../types';
 import { formatTask } from '../utils/taskUtils';
 
 
@@ -38,6 +38,10 @@ export class TaskTreeItem extends TreeItem {
 				this.iconPath = new ThemeIcon('pass', new ThemeColor('todomd.treeViewCompletedTaskIcon'));
 			}
 		}
+
+		if (task.due?.isDue === IsDue.NotDue) {
+			this.description = task.due?.closestDueDateInTheFuture;
+		}
 	}
 }
 
@@ -58,7 +62,7 @@ export class TaskProvider implements TreeDataProvider<TaskTreeItem> {
 	 * Resolve `tooltip` only on hover
 	 */
 	resolveTreeItem(item: TaskTreeItem, el: TaskTreeItem) {
-		el.tooltip = getTaskHoverMd(el.task);
+		el.tooltip = getTasksHoverMd([el.task]);
 		return el;
 	}